refactor: file upload module encapsulation
